backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
python-multipart==0.0.6
openpyxl==3.1.2
motor==3.7.1
pydantic==1.10.11
python-dotenv==1.0.1
pymongo==4.5.0